requirements.txt
# Core PyTorch
# Asegúrate de instalar la variante de CUDA adecuada para tu sistema si usas GPU
# Ejemplo: pip install torch==2.3.1+cu121 -f https://download.pytorch.org/whl/torch_stable.html
torch>=2.0.0

# Procesamiento de datos
numpy>=1.21.0
pandas>=1.4.0,<3.0
scikit-learn>=1.2.0
scipy>=1.9.0

# Visualización (opcional pero recomendado)
matplotlib>=3.5.0
seaborn>=0.11.0

# Monitoreo y logging (opcional)
wandb>=0.15.0

# Utilidades
tqdm>=4.64.0
PyYAML>=6.0
optuna>=3.6.0

# Testing (opcional para desarrollo)
pytest>=7.0.0
pytest-cov>=4.0.0

# Paquetes Financieros
alpha_vantage>=2.3.1
yfinance>=0.2.18
pandas-ta-classic>=0.3.14b,<0.4

# Gestión de variables de entorno
python-dotenv>=1.0.0
